Split backwards-compatibility.test.ts into another function helpers.ts
