+++ src/tests/backwards-compatibility/backwards-compatibility.test.ts
@@ -3,16 +3,23 @@ import * as path from 'path';
 
 import {
   ApiClass,
-  ApiConstructor,
   ApiEnum,
   ApiEnumMember,
   ApiFunction,
-  ApiItem,
-  ApiModel,
   ApiProperty,
   ApiTypeAlias,
-  Parameter
 } from '@microsoft/api-extractor-model';
+
+import {
+  loadApiData,
+  getFunctions,
+  getConstructor,
+  getEnums,
+  getClasses,
+  getProperties,
+  getTypes,
+  checkFunctionCompatibility,
+} from './helpers';
 
 describe('Backwards Compatibility', () => {
   let failure = false;
@@ -29,44 +36,6 @@ describe('Backwards Compatibility', () => {
     }
   });
 
-  // Helper function to load API data
-  const loadApiData = (): { newApiMembers: readonly ApiItem[], currentApiMembers: readonly ApiItem[] } => {
-    if (!fs.existsSync(newApiJsonPath)) {
-      throw new Error(
-        'API reports not found. Run the generate-api-report test first.'
-      );
-    }
-
-
-    const newApiModel = new ApiModel().loadPackage(newApiJsonPath);
-    const newApiMembers = newApiModel.entryPoints[0].members;
-
-    const currentApiModel = new ApiModel().loadPackage(currentApiJsonPath);
-    const currentApiMembers = currentApiModel.entryPoints[0].members;
-
-    return { newApiMembers, currentApiMembers };
-  };
-
-  // Helper functions for getting different kinds of items from the API members
-  const getFunctions = (members: readonly ApiItem[]): ApiFunction[] => {
-    return members.filter((m: ApiItem) => m instanceof ApiFunction && m.kind === 'Function') as ApiFunction[];
-  }
-  const getConstructor = (members: readonly ApiItem[]): ApiConstructor => {
-    return (members.filter((m: ApiItem) => m instanceof ApiConstructor && m.kind === 'Constructor') as ApiConstructor[])[0];
-  }
-  const getEnums = (members: readonly ApiItem[]): ApiEnum[] => {
-    return members.filter((m: ApiItem) => m instanceof ApiEnum && m.kind === 'Enum') as ApiEnum[];
-  }
-  const getClasses = (members: readonly ApiItem[]): ApiClass[] => {
-    return members.filter((m: ApiItem) => m instanceof ApiClass && m.kind === 'Class') as ApiClass[];
-  }
-  const getProperties = (members: readonly ApiItem[]): ApiProperty[] => {
-    return members.filter((m: ApiItem) => m instanceof ApiProperty && m.kind === 'Property') as ApiProperty[];
-  }
-  const getTypes = (members: readonly ApiItem[]): ApiTypeAlias[] => {
-    return members.filter((m: ApiItem) => m instanceof ApiTypeAlias && m.kind === 'TypeAlias') as ApiTypeAlias[];
-  }
-
   describe('Exports', () => {
     describe('should verify that all exports in current are still in new', () => {
       const { newApiMembers, currentApiMembers } = loadApiData();
@@ -78,47 +47,6 @@ describe('Backwards Compatibility', () => {
       });
     });
   });
-
-  const checkFunctionCompatibility = (newFunction: ApiFunction | ApiConstructor, currentFunction: ApiFunction | ApiConstructor) => {
-    const lengthOfPreviousParameters = currentFunction.parameters.length;
-
-    it(`Function ${newFunction.displayName} should have at least as many parameters as the current function`, () => {
-      expect(newFunction.parameters.length).toBeGreaterThanOrEqual(currentFunction.parameters.length);
-    });
-
-    it(`Function ${newFunction.displayName} should have parameters in the same order as the current function`, () => {
-      const newFunctionParamNames = newFunction.parameters.slice(0, lengthOfPreviousParameters).map((p: Parameter) => p.name);
-      const currentFunctionParamNames = currentFunction.parameters.map((p: Parameter) => p.name);
-      expect(newFunctionParamNames).toEqual(currentFunctionParamNames);
-    });
-
-    it(`Function ${newFunction.displayName} should have compatible parameter types with the current function`, () => {
-      const newFunctionParamTypes = newFunction.parameters.slice(0, lengthOfPreviousParameters).map((p: Parameter) => p.parameterTypeExcerpt.text);
-      const currentFunctionParameterTypes = currentFunction.parameters.map((p: Parameter) => p.parameterTypeExcerpt.text);
-      expect(newFunctionParamTypes).toEqual(currentFunctionParameterTypes);
-    });
-    
-    // Check return type compatibility
-    // This check fails if it's a constructor, as those don't have a return type
-    if(currentFunction instanceof ApiFunction && newFunction instanceof ApiFunction){
-      if(!currentFunction.returnTypeExcerpt?.isEmpty) {
-        it(`Function ${newFunction.displayName} should have the same return type as the current function`, () => {
-          expect(newFunction.returnTypeExcerpt.text).toEqual(currentFunction.returnTypeExcerpt.text);
-        });
-      }
-    }
-
-    it(`Function ${newFunction.displayName} should have all new parameters as optional`, () => {
-      const newParameters = newFunction.parameters.slice(lengthOfPreviousParameters);
-      expect(newParameters.every((p: Parameter) => p.isOptional)).toBe(true);
-    });
-
-    it(`Function ${newFunction.displayName} should not have any optional parameters that became required`, () => {
-      const requiredParameters = newFunction.parameters.filter((p: Parameter) => !p.isOptional);
-      const currentRequiredParameters = currentFunction.parameters.filter((p: Parameter) => !p.isOptional);
-      expect(requiredParameters.length).toBeLessThanOrEqual(currentRequiredParameters.length);
-    });
-  }
 
   describe('Functions', () => {
     it('should have at least as many functions in new API as in current', () => {
